ai-chat-api/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.4.2
pydantic-settings>=2.8.1
boto3>=1.28.0
opensearch-py[async]>=2.3.0
pytest>=7.4.2